Bug 1467445 - Revert ErrorAggregator refactor
Recent refactor of ErrorAggregator and ExperimentsErrorAggregator (156089a) is causing runtime issues.
This commit is reverting changes in above mentioned classes.
